Correct tracking maintenance parts
2025-09-05

- change: fix codesmells
- change: correct generateNewMaintenanceParts() in MalfunctionManager

+++ mars-sim-core/src/main/java/com/mars_sim/core/goods/GoodsManager.java
@@ -827,8 +827,8 @@ public class GoodsManager implements Serializable {
 		int optimalPerPop = limits.optimal();
 		int pop = settlement.getNumCitizens();
 		
-		double optimal = optimalPerPop * pop;
-		double reserve = reservePerPop * pop;
+		int optimal = optimalPerPop * pop;
+		int reserve = reservePerPop * pop;
 		double demand = getDemandScoreWithID(resourceID);
 
 		// Compare the available amount of oxygen
